Extract ProductSummaryImage
